PP-284: HDBT syncing
