+++ conf/cmi/block.block.helfi_paatokset_eucookiecomplianceblock.yml
@@ -10,7 +10,7 @@ dependencies:
 id: helfi_paatokset_eucookiecomplianceblock
 theme: helfi_paatokset
 region: content
-weight: -15
+weight: -16
 provider: null
 plugin: eu_cookie_compliance_block
 settings:
