+++ conf/cmi/block.block.helfi_paatokset_heroblock.yml
@@ -3,14 +3,14 @@ langcode: en
 status: true
 dependencies:
   module:
-    - ctools
     - hdbt_content
+    - node
   theme:
     - helfi_paatokset
 id: helfi_paatokset_heroblock
 theme: helfi_paatokset
 region: before_content
-weight: -12
+weight: -15
 provider: null
 plugin: hero_block
 settings:
