+++ conf/cmi/block.block.views_block__declarations_of_affiliation_block_1.yml
@@ -11,8 +11,8 @@ dependencies:
     - helfi_paatokset
 id: views_block__declarations_of_affiliation_block_1
 theme: helfi_paatokset
-region: after_content
-weight: -9
+region: content
+weight: -13
 provider: null
 plugin: 'views_block:declarations_of_affiliation-block_1'
 settings:
