+++ public/modules/custom/paatokset_policymakers/paatokset_policymakers.module
@@ -27,6 +27,9 @@ function paatokset_policymakers_theme($existing, $type, $theme, $path) {
     'policymaker_minutes' => [
       'render element' => 'elements',
       'template' => 'content/minutes',
+      'variables' => [
+        'page_has_sidebar' => true
+      ]
     ],
     'block__policymaker_members' => [
       'base hook' => 'entity',
@@ -209,5 +212,24 @@ function paatokset_policymakers_pathauto_pattern_alter(&$pattern, array $context
   if (!$node->get('field_dm_org_name')->isEmpty()) {
     $pattern->setPattern('paattajat/[node:field_ahjo_title]-[node:field_dm_org_name:value]');
     return;
+  }
+}
+
+function paatokset_policymakers_preprocess_html(&$variables) {
+  if(isset($variables['page']['sidebar_first']) && array_key_exists('policymakersidenavigation', $variables['page']['sidebar_first'])) {
+    $variables['attributes']['class'][] = 'page-type--policymaker';
+  }
+}
+
+/**
+ * Set a variable to always show sidebar in policymaker pages.
+ * 
+ * Implements hook_preprocess_page().
+ */
+function paatokset_policymakers_preprocess_page(&$variables) {
+  if(isset($variables['page']['sidebar_first']) && array_key_exists('policymakersidenavigation', $variables['page']['sidebar_first'])) {
+    $variables['page_has_sidebar'] = true;
+    $variables['move_before_content'] = true;
+    $variables['has_navigation'] = true;
   }
 }
