+++ public/modules/custom/paatokset_policymakers/templates/block/block--policymaker-members.html.twig
@@ -1,7 +1,6 @@
 {%
   set classes = [
   'policymaker-members__block',
-  'container',
   view_mode ? 'policymaker-members--view-mode-' ~ view_mode|clean_class,
 ]
 %}
