+++ public/modules/custom/paatokset_submenus/templates/block--policymaker-side-navigation--mobile.html.twig
@@ -1,0 +1,6 @@
+
+{% if dropdown_attributes.options %}
+  <div class="policymaker-side-menu policymaker-side-menu--mobile">
+    <hds-dropdown{{ dropdown_attributes }} ></hds-dropdown>
+  </div>
+{% endif %}
