+++ public/modules/custom/paatokset_submenus/templates/block--policymaker-side-navigation.html.twig
@@ -81,9 +81,3 @@
     </div>
   {% endif %}
 {% endmacro %}
-
-{% if dropdown_attributes.options %}
-  <div class="policymaker-side-menu policymaker-side-menu--mobile">
-    <hds-dropdown{{ dropdown_attributes }} ></hds-dropdown>
-  </div>
-{% endif %}
