+++ public/themes/custom/helfi_paatokset/src/scss/components/block/_policymaker_listing.scss
@@ -3,7 +3,8 @@
   background-color: #f7f7f8;
 
   .container {
-    padding: 65px 0;
+    padding-top: 65px;
+    padding-bottom: 65px;
   }
 
   h2 {
@@ -55,6 +56,10 @@
   .policymaker-card__title-icon {
     padding: 25px;
     background-color: $color-white;
+
+    h4 {
+      margin-top: 0;
+    }
   }
 
   .hel-icon {
