+++ public/themes/custom/helfi_paatokset/src/scss/components/block/_title.scss
@@ -1,3 +1,7 @@
-.page-title {
-  margin-top: 0;
+.page-type--policymaker {
+  &.has-sidebar, .has-sidebar {
+    .before-content--in-main .page-title {
+      margin-top: 0;
+    }
+  }
 }
