+++ public/themes/custom/helfi_paatokset/src/scss/components/navigation/_menu.scss
@@ -67,3 +67,7 @@
     }
   }
 }
+
+.header-bottom .menu__toggle-button {
+  background-color: transparent;
+}
