--- public/themes/custom/helfi_paatokset/src/scss/layout/_before-content.scss
@@ -1,3 +1,0 @@
-.before-content {
-  width: 100%;
-}
