--- public/themes/custom/helfi_paatokset/src/scss/layout/_container.scss
@@ -1,37 +1,0 @@
-//
-// @file
-//
-// This file includes styles for a container.
-//
-.container {
-  margin-left: auto;
-  margin-right: auto;
-  padding-left: $spacing-s;
-  padding-right: $spacing-s;
-  padding-top: 0;
-  transition: padding 150ms linear;
-
-  max-width: 100%;
-
-  @include breakpoint($breakpoint-xs) {
-    max-width: $container-width-xs;
-    padding-left: 0;
-    padding-right: 0;
-  }
-
-  @include breakpoint($breakpoint-s) {
-    max-width: $container-width-s;
-  }
-  
-  @include breakpoint($breakpoint-m) {
-    max-width: $container-width-m;
-  }
-
-  @include breakpoint($breakpoint-l) {
-    max-width: $container-width-l;
-  }
-
-  @include breakpoint($breakpoint-xl) {
-    max-width: $container-width-xl;
-  }
-}
