--- public/themes/custom/helfi_paatokset/src/scss/layout/_layout-content.scss
@@ -1,28 +1,0 @@
-.main-content {
-  margin-top: $spacing-double + 8px;
-}
-
-.layout-content {
-  margin: 0 $spacing-double;
-
-  // Negate layout content margins on articles since we want node content to be able to strech from edge to edge.
-  > article,
-  > .page-title {
-    margin: 0 (-$spacing-double);
-  }
-}
-
-.has-sidebar--first {
-  .layout-content {
-    @include breakpoint($breakpoint-l) {
-      margin: 0;
-      order: 2;
-      width: 100%;
-
-      > article,
-      > .page-title {
-        margin: 0;
-      }
-    }
-  }
-}
