--- public/themes/custom/helfi_paatokset/src/scss/layout/_layout-main-wrapper.scss
@@ -1,13 +1,0 @@
-.has-sidebar--first .layout-main-wrapper {
-  @include breakpoint($breakpoint-l) {
-    display: flex;
-    flex-wrap: nowrap;
-    margin: $spacing-and-half auto 0;
-    max-width: $container-width-l;
-  }
-
-  @include breakpoint($breakpoint-xl) {
-    margin-top: 97px;
-    max-width: $container-width-xl;
-  }
-}
