--- public/themes/custom/helfi_paatokset/src/scss/layout/_layout-sidebar-first.scss
@@ -1,40 +1,0 @@
-.has-sidebar .layout-sidebar-first,
-.layout-sidebar-first {
-  @include breakpoint($breakpoint-l) {
-    width: 30%;
-    max-width: 308px;
-    order: 0;
-    margin-right: $spacing-and-half;
-  }
-
-  @include breakpoint($breakpoint-xl) {
-    margin-right: 97px;
-  }
-
-  .sidebar-first__container {
-    @include breakpoint($breakpoint-l) {
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: flex-end;
-      padding-left: 0;
-      margin-top: 0;
-    }
-  }
-
-  .container.sidebar-first__container {
-    width: 100%;
-    float: none;
-  }
-
-  .sidebar-first {
-    width: 100%;
-  
-    @include breakpoint($breakpoint-l) {
-      padding-left: 0;
-    }
-
-    @include breakpoint($breakpoint-xl) {
-      padding-left: 0;
-    }
-  }
-}
