--- public/themes/custom/helfi_paatokset/src/scss/layout/_main-content.scss
@@ -1,9 +1,0 @@
-.has-sidebar--first .main-content {
-  @include breakpoint($breakpoint-l) {
-    display: block;
-    max-width: $content-area-wrapper-width-max;
-    margin-top: 0;
-    padding: 0;
-    width: 100%;
-  }
-}
