--- public/themes/custom/helfi_paatokset/src/scss/layout/_page.scss
@@ -1,35 +1,0 @@
-//
-// @file
-//
-// This file includes styles for a page layout.
-//
-.page-content {
-  padding: $spacing-s 0;
-
-  @include breakpoint($breakpoint-l) {
-    padding: $spacing-xl 0;
-  }
-}
-
-.content__wrapper {
-  display: block;
-
-  @include breakpoint($breakpoint-l) {
-    display: flex;
-  }
-}
-
-.container.sidebar-first__container {
-  @include breakpoint($breakpoint-l) {
-    margin-right: $spacing-l;
-    max-width: 270px;
-    min-width: 270px;
-    padding: 35px 0 0 0;
-  }
-
-  @include breakpoint($breakpoint-xl) {
-    max-width: 300px;
-    min-width: 300px;
-  }
-  float:left;
-}
