+++ public/themes/custom/helfi_paatokset/src/scss/layout/_paragraph-content.scss
@@ -1,4 +1,4 @@
-.paragraph_content {
+.paragraph-content {
   > *:last-child {
     margin-bottom: 0;
   }
