+++ public/themes/custom/helfi_paatokset/src/scss/layout/_paragraph-news-liftups.scss
@@ -62,6 +62,7 @@
     h3 {
       font-size: 1rem;
       margin-bottom: $spacing-half;
+      margin-top: 0;
       font-weight: $font-weight-medium
     }
