--- public/themes/custom/helfi_paatokset/src/scss/layout/_paragraph_hero.scss
@@ -1,9 +1,0 @@
-.hero--full {
-  margin-left: -$spacing-double;
-  margin-right: -$spacing-double;
-
-  .container.hero__Container {
-    padding-left: $spacing-double;
-    padding-right: $spacing-double;
-  }
-}
