--- public/themes/custom/helfi_paatokset/src/scss/layout/_sidebar-first.scss
@@ -1,3 +1,0 @@
-.sidebar-first__container {
-  margin-top: 32px;
-}
