--- public/themes/custom/helfi_paatokset/templates/layout/page.html.twig
@@ -1,103 +1,0 @@
-{#
-/**
- * @file
- * Theme override to display a single page.
- */
-#}
-
-{% if page.sidebar_first|render|striptags is not empty %}
-  {% set sidebar_first_output = TRUE %}
-{% endif %}
-
-{% if page.sidebar_second|render|striptags is not empty %}
-  {% set sidebar_second_output = render_var(page.sidebar_second) %}
-{% endif %}
-
-{% set page_classes = [
-    node.getType() ? 'page-type--' ~ node.getType()|clean_class,
-    'wrapper',
-    sidebar_first_output ? 'has-sidebar has-sidebar--first',
-    sidebar_second_output ? 'has-sidebar has-sidebar--second'
-  ]
-%}
-
-<div{{ attributes.addClass(page_classes) }}>
-  {% if page.header_top or page.header_bottom or page.header_branding %}
-    <header role="banner" class="header">
-      {% if page.header_top %}
-        {{ page.header_top }}
-      {% endif %}
-
-      {% if page.header_branding %}
-        {{ page.header_branding }}
-      {% endif %}
-
-      {% if page.header_bottom %}
-        {{ page.header_bottom }}
-      {% endif %}
-    </header>
-  {% endif %}
-
-  {% if page.messages %}
-    {{ page.messages }}
-  {% endif %}
-
-  {% if page.breadcrumb %}
-    {{ page.breadcrumb }}
-  {% endif %}
-
-  {% if page.tools %}
-    {{ page.tools }}
-  {% endif %}
-
-  <main role="main" class="layout-main-wrapper">
-    {% if sidebar_first_output %}
-      <aside class="layout-sidebar-first" role="complementary">
-        {{ page.sidebar_first }}
-      </aside>
-    {% endif %}
-
-    <a id="main-content" tabindex="-1"></a>
-    <div class="main-content">
-      <div class="layout-content">
-        {% if page.before_content %}
-          {{ page.before_content }}
-        {% endif %}
-
-        {{ page.content }}
-
-        {% if page.after_content %}
-          {{ page.after_content }}
-        {% endif %}
-      </div>
-    </div>
-
-    {% if sidebar_second_output %}
-      <aside class="layout-sidebar-second" role="complementary">
-        {{ page.sidebar_second }}
-      </aside>
-    {% endif %}
-  </main>
-</div>
-
-{% if page.footer_top or page.footer_bottom %}
-  {% set footer_variant = '' %}
-  {% if page.footer_color %}
-    {% set footer_variant = 'footer--' ~ page.footer_color|clean_class %}
-  {% endif %}
-
-  <footer role="contentinfo" class="footer {{ footer_variant }}">
-    {% include '@hdbt/misc/koro.twig' with {koro: koro, flip: false } %}
-    {% if page.footer_top %}
-      {{ page.footer_top }}
-    {% endif %}
-
-    {% if page.footer_bottom %}
-      {{ page.footer_bottom }}
-    {% endif %}
-  </footer>
-{% endif %}
-
-{% if page.attachments %}
-  {{ page.attachments }}
-{% endif %}
